Move check return cycle into jobs folder
It is now only called by the return logs job so it makes sense to move it there rather than leave it in `/services/return-logs` which might give the impression it is used by the services there.
