+++ test/services/jobs/return-logs/check-return-cycle.service.test.js
@@ -9,12 +9,12 @@ const { describe, it, before, beforeEach, afterEach } = (exports.lab = Lab.scrip
 const { expect } = Code
 
 // Things to stub
-const ReturnCycleModel = require('../../../app/models/return-cycle.model.js')
+const ReturnCycleModel = require('../../../../app/models/return-cycle.model.js')
 
 // Thing under test
-const CheckReturnCycleService = require('../../../app/services/return-logs/check-return-cycle.service.js')
+const CheckReturnCycleService = require('../../../../app/services/jobs/return-logs/check-return-cycle.service.js')
 
-describe('Return Logs - Check Return Cycle service', () => {
+describe('Jobs - Return Logs - Check Return Cycle service', () => {
   const changeDate = new Date('2024-05-01')
   let clock
   let insertStub
